chore: move api config to auth server #1621
