+++ admin-ui/plugins/admin/plugin-metadata.js
@@ -15,14 +15,12 @@ import apiRoleSaga from './redux/sagas/ApiRoleSaga'
 import apiPermissionSaga from './redux/sagas/ApiPermissionSaga'
 import mappingSaga from './redux/sagas/MappingSaga'
 import webhookSaga from './redux/sagas/WebhookSaga'
-import configApiSaga from './redux/sagas/configApiSaga'
 
 import { reducer as scriptReducer } from 'Plugins/admin/redux/features/customScriptSlice'
 import { reducer as apiRoleReducer } from 'Plugins/admin/redux/features/apiRoleSlice'
 import { reducer as apiPermissionReducer } from 'Plugins/admin/redux/features/apiPermissionSlice'
 import { reducer as mappingReducer } from 'Plugins/admin/redux/features/mappingSlice'
 import webhookReducer from 'Plugins/admin/redux/features/WebhookSlice'
-import configApiReducer from 'Plugins/admin/redux/features/configApiSlice'
 import {
   ACR_READ,
   ROLE_READ,
@@ -32,11 +30,9 @@ import {
   MAPPING_READ,
   WEBHOOK_READ,
   WEBHOOK_WRITE,
-  API_CONFIG_READ
 } from 'Utils/PermChecker'
 import WebhookAddPage from './components/Webhook/WebhookAddPage'
 import WebhookEditPage from './components/Webhook/WebhookEditPage'
-import ConfigApiPage from './components/Configuration/ConfigApiPage'
 
 const PLUGIN_BASE_APTH = '/adm'
 
@@ -63,11 +59,6 @@ const pluginMetadata = {
               title: 'menus.api.mapping',
               path: PLUGIN_BASE_APTH + '/mapping',
               permission: MAPPING_READ,
-            },
-            {
-              title: 'menus.api.api_config',
-              path: PLUGIN_BASE_APTH + '/config-api-configuration',
-              permission: API_CONFIG_READ,
             },
           ],
         },
@@ -160,11 +151,6 @@ const pluginMetadata = {
       path: PLUGIN_BASE_APTH + '/webhook/edit/:id',
       permission: WEBHOOK_WRITE,
     },
-    {
-      component: ConfigApiPage,
-      path: PLUGIN_BASE_APTH + '/config-api-configuration',
-      permission: API_CONFIG_READ,
-    },
   ],
   reducers: [
     { name: 'scriptReducer', reducer: scriptReducer },
@@ -172,9 +158,8 @@ const pluginMetadata = {
     { name: 'apiPermissionReducer', reducer: apiPermissionReducer },
     { name: 'mappingReducer', reducer: mappingReducer },
     { name: 'webhookReducer', reducer: webhookReducer },
-    { name: 'configApiReducer', reducer: configApiReducer },
   ],
-  sagas: [scriptSaga(), apiRoleSaga(), apiPermissionSaga(), mappingSaga(), webhookSaga(), configApiSaga()],
+  sagas: [scriptSaga(), apiRoleSaga(), apiPermissionSaga(), mappingSaga(), webhookSaga()],
 }
 
 export default pluginMetadata
