+++ admin-ui/plugins/auth-server/redux/sagas/configApiSaga.js
@@ -2,10 +2,10 @@ import { all, call, fork, put, select, takeLatest } from 'redux-saga/effects'
 import { initAudit } from 'Redux/sagas/SagaUtils'
 const JansConfigApi = require('jans_config_api')
 import { getClient } from 'Redux/api/base'
-import ConfigurationApi from '../api/ConfigurationApi'
+import ConfigurationApi from 'Plugins/auth-server/redux/api/ConfigurationApi.js'
 import { isFourZeroOneError, addAdditionalData } from 'Utils/TokenController'
 import { FETCH, PATCH } from '../../../../app/audit/UserActionType'
-import { getConfigApiConfigurationResponse } from '../features/configApiSlice'
+import { getConfigApiConfigurationResponse } from 'Plugins/auth-server/redux/features/configApiSlice.js'
 import { postUserAction } from 'Redux/api/backend-api'
 import { updateToast } from 'Redux/features/toastSlice'
 import { getAPIAccessToken } from 'Redux/features/authSlice'
